Split and move TraceExporter to a different package. Use the same idea of component.
