+++ api/src/test/java/io/opencensus/trace/TraceComponentTest.java
@@ -17,6 +17,7 @@ import static com.google.common.truth.Truth.assertThat;
 
 import io.opencensus.internal.ZeroTimeClock;
 import io.opencensus.trace.config.TraceConfig;
+import io.opencensus.trace.export.ExportComponent;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -43,7 +44,7 @@ public class TraceComponentTest {
   @Test
   public void defaultTraceExporter() {
     assertThat(TraceComponent.getNoopTraceComponent().getTraceExporter())
-        .isSameAs(TraceExporter.getNoopTraceExporter());
+        .isSameAs(ExportComponent.getNoopExportComponent());
   }
 
   @Test
