+++ core_impl/src/test/java/io/opencensus/trace/SpanExporterImplTest.java
@@ -22,7 +22,7 @@ import io.opencensus.common.SimpleEventQueue;
 import io.opencensus.trace.Span.Options;
 import io.opencensus.trace.SpanImpl.StartEndHandler;
 import io.opencensus.trace.config.TraceParams;
-import io.opencensus.trace.TraceExporter.SpanExporter.Handler;
+import io.opencensus.trace.export.SpanExporter.Handler;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.EnumSet;
@@ -37,7 +37,7 @@ import org.junit.runners.JUnit4;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-/** Unit tests for {@link TraceExporterImpl}. */
+/** Unit tests for {@link ExportComponentImpl}. */
 @RunWith(JUnit4.class)
 public class SpanExporterImplTest {
   private static final String SPAN_NAME_1 = "MySpanName/1";
@@ -51,8 +51,7 @@ public class SpanExporterImplTest {
   private final SpanContext notSampledSpanContext =
       SpanContext.create(
           TraceId.generateRandomId(random), SpanId.generateRandomId(random), TraceOptions.DEFAULT);
-  private final SpanExporterImpl sampledSpansServiceExporter =
-      SpanExporterImpl.create(4, 1000);
+  private final SpanExporterImpl sampledSpansServiceExporter = SpanExporterImpl.create(4, 1000);
   private final StartEndHandler startEndHandler =
       new StartEndHandlerImpl(sampledSpansServiceExporter, new SimpleEventQueue());
   private EnumSet<Options> recordSpanOptions = EnumSet.of(Options.RECORD_EVENTS);
