+++ core_impl_java/src/test/java/io/opencensus/trace/TracingTest.java
@@ -41,6 +41,6 @@ public class TracingTest {
 
   @Test
   public void implementationOfTraceExporter() {
-    assertThat(Tracing.getTraceExporter()).isInstanceOf(TraceExporterImpl.class);
+    assertThat(Tracing.getTraceExporter()).isInstanceOf(ExportComponentImpl.class);
   }
 }
